chore: separate homepage state hook from context
