+++ packages/apps/human-app/frontend/src/modules/homepage/components/choose-sign-up-account-type.tsx
@@ -7,8 +7,8 @@ import { routerPaths } from '@/router/router-paths';
 import { PageCard } from '@/shared/components/ui/page-card';
 import type { HomePageStageType } from '@/modules/homepage/views/home.page';
 import { useColorMode } from '@/shared/hooks/use-color-mode';
-import { useHomePageState } from '@/shared/contexts/homepage-state';
 import { onlyDarkModeColor } from '@/shared/styles/dark-color-palette';
+import { useHomePageState } from '@/shared/hooks/use-homepage-state';
 
 interface ChooseSignUpAccountType {
   setStage: (step: HomePageStageType) => void;
