+++ packages/apps/human-app/frontend/src/modules/homepage/views/home.page.tsx
@@ -8,8 +8,8 @@ import { useAuth } from '@/modules/auth/hooks/use-auth';
 import { routerPaths } from '@/router/router-paths';
 import { Button } from '@/shared/components/ui/button';
 import { useColorMode } from '@/shared/hooks/use-color-mode';
-import { useHomePageState } from '@/shared/contexts/homepage-state';
 import { HomeContainer } from '@/modules/homepage/components/home-container';
+import { useHomePageState } from '@/shared/hooks/use-homepage-state';
 
 export type HomePageStageType = 'welcome' | 'chooseSignUpAccountType';
