+++ packages/apps/human-app/frontend/src/shared/hooks/use-homepage-state.tsx
@@ -1,0 +1,12 @@
+import { useContext } from 'react';
+import { HomePageStateContext } from '../contexts/homepage-state';
+
+export const useHomePageState = () => {
+  const context = useContext(HomePageStateContext);
+  if (!context) {
+    throw new Error(
+      'useHomePageState must be used within a HomePageStageProvider'
+    );
+  }
+  return context;
+};
