+++ packages/apps/human-app/frontend/src/shared/types/homepage-state.ts
@@ -1,0 +1,1 @@
+export type HomePageStateType = 'welcome' | 'chooseSignUpAccountType';
